fix: updating manual trace - Jaeger Tracing
fixing OpenBLAS Warning
