+++ .env
@@ -21,6 +21,9 @@ GRAFANA_IMAGE_TAG="11.3.4"
 # jaeger images & containers
 JAEGER_CONTAINER_NAME="jaeger"
 JAEGER_PORT=16686
+JAEGER_PORT_UDP=6831
 JAEGER_IMAGE_TAG="1.66.0"
+COLLECTOR_ZIPKIN_HOST_PORT=9411
+JAEGER_COLLECTOR_PORT=14268
 OTEL_EXPORTER_OTLP_ENDPOINT="http://jaeger:4317"
 OTEL_TRACES_EXPORTER="otlp"
