+++ rag-pipeline/coverage.xml
@@ -1,5 +1,5 @@
 <?xml version="1.0" ?>
-<coverage version="7.7.1" timestamp="1742773441942" lines-valid="250" lines-covered="212" line-rate="0.848" branches-covered="0" branches-valid="0" branch-rate="0" complexity="0">
+<coverage version="7.7.1" timestamp="1742776847750" lines-valid="250" lines-covered="212" line-rate="0.848" branches-covered="0" branches-valid="0" branch-rate="0" complexity="0">
 	<!-- Generated by coverage.py: https://coverage.readthedocs.io/en/7.7.1 -->
 	<!-- Based on https://raw.githubusercontent.com/cobertura/web/master/htdocs/xml/coverage-04.dtd -->
 	<sources>
