+++ rag-pipeline/requirements.txt
@@ -25,3 +25,4 @@ opentelemetry-sdk==1.31.0
 opentelemetry-exporter-otlp==1.31.0
 opentelemetry-instrumentation-fastapi==0.52b0
 opentelemetry-instrumentation-requests==0.52b0
+opentelemetry-exporter-jaeger==1.21.0
